====HACK===: x86/ioapic: Add support for IRQCHIP_CAN_DELIVER_AS_NMI
This is a hack. Setting the delivery mode of the redirection entry
as NMI works. However, this behavior is undocumented.

The only problem is that if NMI is used as delivery mode, there are
problems when setting the destination ID. irq_set_affinity() does not
work, all interrupts are delivered to a single CPU.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>